% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{add_metadata}
\alias{add_metadata}
\title{Add metadata to the JECFA dataset}
\usage{
add_metadata(jecfa)
}
\arguments{
\item{jecfa}{A data frame with the JECFA data}
}
\value{
A data frame with the JECFA data and the added metadata
}
\description{
This function adds metadata to the JECFA dataset. The metadata
includes the type of the JECFA report, the host of the URL, the
abbreviation of the JECFA report, the clean report, the FAS number,
and a reference ID.
}
\examples{
c(1, 2) |>
  purrr::map(\(id) compose_jecfa_list(id) |> get_result()) |>
  create_df() |>
  process_df() |>
  add_metadata()
}
