% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_filepaths.R
\name{compose_filepaths}
\alias{compose_filepaths}
\title{Compose filepaths to download TRS}
\usage{
compose_filepaths(urls, dir, noid = FALSE)
}
\arguments{
\item{urls}{(chr) URLs to download}

\item{dir}{(chr) Directory to save the files}

\item{noid}{(logical) Whether to use the noid version of the filename}
}
\value{
(chr) Filepaths to download
}
\description{
Compose filepaths to download TRS
}
\examples{
c(1, 10) |>
  purrr::map(\(id) compose_jecfa_list(id) |> get_result()) |>
  create_df() |>
  process_df() |>
  add_metadata() |>
  compose_urls() |>
  remove_null_urls() |>
  compose_filepaths(tempdir())
}
