% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_jecfa_list.R
\name{compose_jecfa_list}
\alias{compose_jecfa_list}
\title{Compose JECFA list}
\usage{
compose_jecfa_list(jecfa_id)
}
\arguments{
\item{jecfa_id}{(int) The index of the chemical in the JECFA
database.}
}
\value{
A list with two elements: result and error. The result is a
tibble with the chemical information. The error is NULL if the
function runs successfully; otherwise, it contains the error
message, and result is NULL.
}
\description{
This function composes a list of JECFA chemicals.
}
\examples{
compose_jecfa_list(1) # success
compose_jecfa_list("a") # failure
}
