% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_filepaths.R
\name{compose_maptojecfa}
\alias{compose_maptojecfa}
\title{Map TRS path to JECFA}
\usage{
compose_maptojecfa(x)
}
\arguments{
\item{x}{(chr) TRS path}
}
\value{
(tibble) JECFA ID and file
}
\description{
Map TRS path to JECFA
}
\examples{
# only record 10 and 11 has TRS
jecfa_sample <- c(1, 10, 11) |>
  purrr::map(\(id) compose_jecfa_list(id) |> get_result()) |>
  create_df() |>
  process_df() |>
  add_metadata()

 # id 2 and 3 of the provided db (record 10, and 11) has TRS and
 # are reported in the resulting tibble
 jecfa_sample |>
  compose_urls() |>
  remove_null_urls() |>
  compose_filepaths(tempdir()) |>
  compose_maptojecfa()
}
