% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_for_tm.R
\name{compose_tbl10}
\alias{compose_tbl10}
\title{Table 10}
\usage{
compose_tbl10(jecfa_tm_full)
}
\arguments{
\item{jecfa_tm_full}{A tibble with the JECFA data}
}
\value{
A gtsummary table
}
\description{
This function creates a table with the distribution of identifiers
when at least one among CAS, COE and FEMA is present
}
