% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_dataset.R
\name{compose_tbl4}
\alias{compose_tbl4}
\title{Compose table 4}
\usage{
compose_tbl4(jecfa)
}
\arguments{
\item{jecfa}{a data frame with the JECFA data}
}
\value{
a gtsummary table
}
\description{
Create a table with the distribution of strange cases
}
\examples{
jecfa_sample <- c(1, 2) |>
  purrr::map(\(id) compose_jecfa_list(id) |> get_result()) |>
  create_df() |>
  process_df() |>
  add_metadata()

compose_tbl4(jecfa_sample)
}
