% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_for_tm.R
\name{compose_tbl5}
\alias{compose_tbl5}
\title{Table 5}
\usage{
compose_tbl5(keywordMatching)
}
\arguments{
\item{keywordMatching}{A tibble with the results of the keyword
matching}
}
\value{
A gtsummary table
}
\description{
This function creates a table with the distribution of retrieved
words by source document
}
