% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_for_tm.R
\name{compose_tbl6}
\alias{compose_tbl6}
\title{Table 6}
\usage{
compose_tbl6(keywordMatching)
}
\arguments{
\item{keywordMatching}{A tibble with the results of the keyword
matching}
}
\value{
A gtsummary table
}
\description{
This function creates a table with the distribution of retrieved
words by text bmd and document source
}
