% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_for_tm.R
\name{compose_tbl8}
\alias{compose_tbl8}
\title{Table 8}
\usage{
compose_tbl8(jecfa_tm_full)
}
\arguments{
\item{jecfa_tm_full}{A tibble with the JECFA data}
}
\value{
A gtsummary table
}
\description{
This function creates a table with the distribution of retrieved
documents
}
