% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_for_tm.R
\name{compose_tbl9}
\alias{compose_tbl9}
\title{Table 9}
\usage{
compose_tbl9(jecfa_tm_full)
}
\arguments{
\item{jecfa_tm_full}{A tibble with the JECFA data}
}
\value{
A gtsummary table
}
\description{
This function creates a table with the distribution of unique
records depending on the set of variables considered
}
