% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_urls.R
\name{compose_url}
\alias{compose_url}
\title{Compose URL to PDF}
\usage{
compose_url(id, db)
}
\arguments{
\item{id}{JECFA ref_id}

\item{db}{JECFA data frame}
}
\value{
a list with the URL, the filename with the ID and the
filename without the ID
}
\description{
Compose URL to TRS PDF for a JECFA report. Valid URLs are
constructed based on the JECFA data frame information, i.e,
they should report "NOT FAS" or "NOT PREPARED" in the
\code{Tox_monograph_abbr} column and have a valid URL in the
\code{Report_sourcelink} column.
}
\examples{
library(jecfa)

jecfa_sample <- c(1, 10) |>
  purrr::map(\(id) compose_jecfa_list(id) |> get_result()) |>
  create_df() |>
  process_df() |>
  add_metadata()

jecfa_sample[["ref_id"]][[1]] |> # report 1 has no TRS PDF
  compose_url(jecfa_sample)

jecfa_sample[["ref_id"]][[2]] |> # report 10 has a TRS PDF
  compose_url(jecfa_sample)
}
