% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_df.R
\name{create_df}
\alias{create_df}
\title{Create a data frame from a list of data frames}
\usage{
create_df(jecfa_list)
}
\arguments{
\item{jecfa_list}{A list of (JECFA) data frames (as provided by
\code{compose_jecfa_list() |> get_result()})}
}
\value{
A data frame
}
\description{
This function simply creates a data frame from a list of data frames.
}
\examples{
create_df(list(data.frame(a = 1:3), data.frame(b = 4:6)))
}
