% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraxt_jecfa_report_clean.R
\name{extract_jecfa_report_clean}
\alias{extract_jecfa_report_clean}
\title{Extract JECFA report}
\usage{
extract_jecfa_report_clean(report)
}
\arguments{
\item{report}{(chr) The JECFA report code as reported in the JECFA
database}
}
\value{
(chr) The cleaned JECFA report code
}
\description{
This function extracts and clean the JECFA report code.
}
\examples{
compose_jecfa_list(1) |>
  get_result() |>
  dplyr::pull("Report1") |>
  extract_jecfa_report_clean()
}
