% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_jecfa_tox_abbr.R
\name{extract_jecfa_tox_abbr}
\alias{extract_jecfa_tox_abbr}
\title{Extracts JECFA abbreviation}
\usage{
extract_jecfa_tox_abbr(tox_monograph)
}
\arguments{
\item{tox_monograph}{A character vector with the JECFA monograph}
}
\value{
A character vector with the abbreviation of the JECFA
monograph
}
\description{
This function extracts the abbreviation of a JECFA monograph.
}
\examples{
compose_jecfa_list(1) |>
  get_result() |>
  dplyr::pull("Tox Monograph1") |>
  extract_jecfa_tox_abbr()
}
