% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_jecfa_list.R
\name{get_chemical}
\alias{get_chemical}
\title{Get chemical information from JECFA database}
\usage{
get_chemical(id)
}
\arguments{
\item{id}{(chr) The index of the chemical in the JECFA database.}
}
\value{
A tibble with the chemical information.
}
\description{
This function scrapes the JECFA database to get information about a
chemical.
}
\examples{
get_chemical("1")
}
