% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_error}
\alias{get_error}
\title{Get error}
\usage{
get_error(x)
}
\arguments{
\item{x}{(list) A list with a field named 'error'}
}
\value{
The value of the 'error' field
}
\description{
This function extracts the "error" field from a list.
}
\examples{
get_error(list(error = "error message"))
}
