% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_jecfa_list.R
\name{get_jecfa_name}
\alias{get_jecfa_name}
\title{Get JECFA name}
\usage{
get_jecfa_name(page)
}
\arguments{
\item{page}{(xml_document) The page content.}
}
\value{
The name of the chemical.
}
\description{
This function extracts the name of the chemical from the JECFA
database.
}
\examples{
url <- stringr::str_c(
  "https://",
  "apps.who.int/",
  "food-additives-contaminants-jecfa-database/Home/Chemical/",
  "1"
)
page <- rvest::read_html(url)
get_jecfa_name(page)
}
