% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jecfa_type.R
\name{get_jecfa_type}
\alias{get_jecfa_type}
\title{Get the type of JECFA link}
\usage{
get_jecfa_type(tox_link)
}
\arguments{
\item{tox_link}{A character string with the JECFA link}
}
\value{
A character string with the type of JECFA link
}
\description{
Get the type of JECFA link
}
\examples{
compose_jecfa_list(1) |>
  get_result() |>
  dplyr::pull("Tox Monograph_sourcelink1") |>
  get_jecfa_type()
}
