% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jecfa_url_parts.R
\name{get_jecfa_url_parts}
\alias{get_jecfa_url_parts}
\title{Get JECFA URL parts}
\usage{
get_jecfa_url_parts(tox_link)
}
\arguments{
\item{tox_link}{A URL from the JECFA database}
}
\value{
A data frame with the URL parts
}
\description{
Get JECFA URL parts
}
\examples{
compose_jecfa_list(1) |>
  get_result() |>
  dplyr::pull("Tox Monograph_sourcelink1") |>
  get_jecfa_url_parts()

}
