% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_result}
\alias{get_result}
\title{Get result}
\usage{
get_result(x)
}
\arguments{
\item{x}{(list) A list with a field named 'result'}
}
\value{
The value of the 'result' field
}
\description{
This function extracts the "result" field from a list.
}
\examples{
get_result(list(result = 1))
}
