% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jecfa-package.R
\docType{package}
\name{jecfa-package}
\alias{jecfa-package}
\title{jecfa: Web Scraping JECFA}
\description{
Scrapes details related to chemicals from the [Evaluations of the Joint FAO/WHO Expert Committee on Food Additives (JECFA)](https://apps.who.int/food-additives-contaminants-jecfa-database/Home/), captures evaluation reports, links, and other data points.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ubesp-dctv.github.io/jecfa/}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@ubep.unipd.it} (\href{https://orcid.org/0000-0002-0520-7428}{ORCID})

Authors:
\itemize{
  \item Ileana Baldi \email{ileana.baldi@ubep.unipd.it} (\href{https://orcid.org/0000-0002-8578-9164}{ORCID})
}

}
\keyword{internal}
