% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_keywords.R
\name{match_keywords}
\alias{match_keywords}
\title{Match keywords in a parsed PDF}
\usage{
match_keywords(file_path, parsedPdf, keywords, source = c("FAS", "TRS"))
}
\arguments{
\item{file_path}{(chr) The path to the PDF file}

\item{parsedPdf}{(chr) The parsed PDF}

\item{keywords}{(chr) keywords to search for}

\item{source}{(chr) The source of the PDF, either "FAS" or "TRS"}
}
\value{
(tibble) A tibble with the source of the PDF, the file name,
the keywords, the pages where the keywords were found, and whether
any of the keywords were found.
}
\description{
This function takes a parsed PDF and a list of keywords and returns a
tibble with the source of the PDF, the file name, the keywords, the
pages where the keywords were found, and whether any of the keywords
were found.
}
