% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pdf.R
\name{parse_pdf}
\alias{parse_pdf}
\title{Parse PDF}
\usage{
parse_pdf(path, dpi = 150)
}
\arguments{
\item{path}{Path to the PDF file}

\item{dpi}{DPI to use for the OCR}
}
\value{
The text of the PDF file
}
\description{
This function parses a PDF file and returns the text.
}
