% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_df.R
\name{process_df}
\alias{process_df}
\title{Preprocess JECFA data}
\usage{
process_df(df_clean)
}
\arguments{
\item{df_clean}{A data frame with the JECFA data}
}
\value{
A data frame with the cleaned JECFA data
}
\description{
This function cleans the JECFA data by removing unnecessary columns
and renaming the remaining ones.
}
\examples{
c(1, 2) |>
  purrr::map(\(id) compose_jecfa_list(id) |> get_result()) |>
  create_df() |>
  process_df()
}
