% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_null_urls.R
\name{remove_null_urls}
\alias{remove_null_urls}
\title{Remove null TRS URLS}
\usage{
remove_null_urls(x)
}
\arguments{
\item{x}{A list of lists with a url element}
}
\value{
the original lists with a url element filtered to remove nulls
}
\description{
Remove null TRS URLS
}
\examples{
c(1, 10) |>
  purrr::map(\(id) compose_jecfa_list(id) |> get_result()) |>
  create_df() |>
  process_df() |>
  add_metadata() |>
  compose_urls() |>
  remove_null_urls() # only the second element remain, i.e. report 10
}
