% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_urls.R
\name{compose_urls}
\alias{compose_urls}
\title{Compose URLs}
\usage{
compose_urls(jecfa)
}
\arguments{
\item{jecfa}{a data frame with the JECFA data}
}
\value{
a list with the URLs specification for each JECFA report,
i.e., a list with the URL, the filename with the ID and the
filename without the ID
}
\description{
Compose URLs to TRS PDFs for JECFA reports. Valid URLs are
constructed based on the JECFA data frame information, i.e,
they should report "NOT FAS" or "NOT PREPARED" in the
\code{Tox_monograph_abbr} column and have a valid URL in the
\code{Report_sourcelink} column.
}
\examples{
library(jecfa)

# first element will be NULL because report 1 has no PDF, while
# record 2 has a PDF so it will be a list with the URL, the filename
# with the ID and the filename without the ID.
c(1, 10) |>
  purrr::map(\(id) compose_jecfa_list(id) |> get_result()) |>
  create_df() |>
  process_df() |>
  add_metadata() |>
  compose_urls()
}
