% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_distiller_jecfa.R
\name{create_distiller_jecfa}
\alias{create_distiller_jecfa}
\title{Create JECFA distiller}
\usage{
create_distiller_jecfa(jecfa_tm_full)
}
\arguments{
\item{jecfa_tm_full}{A tibble with the JECFA data}
}
\value{
A tibble with the distiller data
}
\description{
This function creates a distiller db from the JECFA data.
}
